Added repo class and moved entity class to it's own package
