+++ src/main/java/com/jn/arcade/ArcadeOwnerApplication.java
@@ -1,4 +1,4 @@
-package com.jn;
+package com.jn.arcade;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
